Add websocket.AuthenticateByRequest
Since browser client on websocket connections won't be able to
send authentication headers, we will need alternative methods that
allows us to authenticate. Using the procedure of authentication by
request will require the client to listen for the AUTH command and
respond with it's credentials which will be used for authentication.
Clients need to implement this.